requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.9.0
bert-tensorflow==1.0.1
boto==2.49.0
boto3==1.12.46
botocore==1.15.46
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
djangorestframework==3.11.0
docutils==0.15.2
gast==0.2.2
gensim==3.8.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.5
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
nltk==3.5
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
praw==6.5.1
prawcore==1.0.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==5.3.1
pyzmq==19.0.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
smart-open==1.11.1
soupsieve==2.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
termcolor==1.1.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
update-checker==0.16
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.0.1
wordcloud==1.6.0
wrapt==1.12.1
